requirements.txt
# Core FastAPI dependencies
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
pydantic-settings==2.2.1

# LlamaExtract dependencies
llama-cloud-services>=0.1.0

# File handling and validation
python-multipart==0.0.6
python-magic==0.4.27
aiofiles==23.2.0

# Background tasks and async support
celery[redis]==5.3.6  # Optional: for production background tasks
redis==5.0.1  # Optional: for caching and task queue

# Security and authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database (optional)
sqlalchemy==2.0.25
alembic==1.13.1
psycopg[binary]==3.2.3
aiosqlite==0.20.0

# Logging and monitoring
structlog==23.2.0

# Development and testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
pytest-cov==4.1.0

# Documentation
mkdocs==1.5.3  # Optional: for documentation generation
mkdocs-material==9.5.6  # Optional: material theme for docs

# Linting and formatting (development)
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0

# S3-compatible storage (DigitalOcean Spaces)
boto3==1.34.162
botocore==1.34.162
